TileEffectManager.cs 리팩토링
모든 확률, 배수 표기 정수 %로 변경

+++ Assets/Scripts/ScriptableObjects/WeaponDatabase.cs
@@ -104,9 +104,9 @@ public class WeaponData
     public int weaponDamage;
     public int weaponRange;
     public int weaponAmmo;
-    public float weaponHitRate;
-    public float weaponCriticalChance;
-    public float weaponCriticalDamage;
+    public int weaponHitRate;
+    public int weaponCriticalChance;
+    public int weaponCriticalDamage;
 
     public int weaponScript;
 }
